PR feedback: Move Sha1DigestKey from base -> rest package (attachment.go)
